requirement.txt
opencv-python-headless
numpy
Pillow
scikit-learn
tensorflow
keras


 